refactor: move levels into subfolder
