+++ frontend/src/game/levels/levelsList/index.ts
@@ -1,9 +1,9 @@
 import testLevel from './testLevel';
-import { LevelConfig } from './levelConfig';
-import { LevelId } from './types';
+import { LevelConfig } from '../levelConfig';
+import LevelId from './levelIds';
 
 const levelList:Record<LevelId, LevelConfig> = {
   [LevelId.test]: testLevel,
 };
 
-export default levelList;
+export { levelList, LevelId };
