+++ frontend/src/game/levels/types/index.ts
@@ -1,8 +1,5 @@
-import LevelId from './levelIds';
 import LevelLoad from './levelLoad';
 import Position from './position';
 import Surface from './surface';
 
-export {
-  LevelId, LevelLoad, Surface, Position,
-};
+export { LevelLoad, Surface, Position };
