+++ frontend/src/game/levels/types/levelLoad.ts
@@ -1,4 +1,4 @@
-import LevelId from './levelIds';
+import LevelId from '../levelsList/levelIds';
 
 type LevelLoad = {
   levelId:LevelId
