+++ frontend/src/game/levels/types/position.ts
@@ -1,3 +1,5 @@
+import { Line } from '../../shapes';
+
 type Position = { position:Line };
 
 export default Position;
